+++ apps/docs/src/components/angular/examples/layout-integration/diagram.component.ts
@@ -3,31 +3,45 @@ import '@angular/compiler';
 import { Component, inject } from '@angular/core';
 import {
   initializeModel,
+  NgDiagramBackgroundComponent,
   NgDiagramComponent,
   NgDiagramModelService,
+  NgDiagramNodeTemplateMap,
   provideNgDiagram,
   type NgDiagramConfig,
   type SelectionMovedEvent,
 } from 'ng-diagram';
 import { diagramModel } from './data';
 import { LayoutButtonsComponent } from './layout-buttons/layout-buttons.component';
+import { NodeComponent } from './node/node.component';
+import { NodeTemplateType } from './types';
 
 @Component({
-  imports: [NgDiagramComponent, LayoutButtonsComponent],
+  imports: [
+    NgDiagramComponent,
+    NgDiagramBackgroundComponent,
+    LayoutButtonsComponent,
+  ],
   template: `
-    <layout-buttons />
     <div class="not-content diagram">
       <ng-diagram
         [model]="model"
         [config]="config"
         (selectionMoved)="onSelectionMoved($event)"
-      />
+        [nodeTemplateMap]="nodeTemplateMap"
+      >
+        <ng-diagram-background />
+      </ng-diagram>
+      <layout-buttons />
     </div>
   `,
-  styleUrls: ['./diagram.component.scss'],
+  styleUrl: './diagram.component.scss',
   providers: [provideNgDiagram()],
 })
 export class DiagramComponent {
+  nodeTemplateMap = new NgDiagramNodeTemplateMap([
+    [NodeTemplateType.CustomNodeType, NodeComponent],
+  ]);
   private modelService = inject(NgDiagramModelService);
 
   model = initializeModel({
@@ -39,6 +53,7 @@ export class DiagramComponent {
     zoom: {
       zoomToFit: {
         onInit: true,
+        padding: 80,
       },
     },
   };