+++ apps/docs/src/components/angular/examples/palette/diagram.component.ts
@@ -3,6 +3,7 @@ import '@angular/compiler';
 import { Component } from '@angular/core';
 import {
   initializeModel,
+  NgDiagramBackgroundComponent,
   NgDiagramComponent,
   provideNgDiagram,
   type NgDiagramConfig,
@@ -11,11 +12,13 @@ import {
 import { Palette } from './palette/palette.component';
 
 @Component({
-  imports: [NgDiagramComponent, Palette],
+  imports: [NgDiagramComponent, NgDiagramBackgroundComponent, Palette],
   providers: [provideNgDiagram()],
   template: `
     <div class="not-content diagram">
-      <ng-diagram [model]="model" [config]="config" />
+      <ng-diagram [model]="model" [config]="config">
+        <ng-diagram-background />
+      </ng-diagram>
     </div>
     <palette-container [model]="paletteModel" />
   `,
@@ -31,7 +34,7 @@ export class DiagramComponent {
     zoom: {
       zoomToFit: {
         onInit: true,
-        padding: [150, 150, 150, 365],
+        padding: [150, 150, 150, 265],
       },
     },
   } satisfies NgDiagramConfig;