+++ apps/docs/src/components/angular/examples/save-state/nav-bar/nav-bar.component.ts
@@ -18,7 +18,7 @@ import { SaveStateService } from '../save.service';
     </div>
   `,
   // @mark-end
-  styleUrls: ['./nav-bar.component.scss'],
+  styleUrl: './nav-bar.component.scss',
 })
 export class NavBarComponent {
   // @collapse-start