+++ apps/docs/src/components/angular/nodes/custom-node/diagram.component.ts
@@ -2,6 +2,7 @@ import '@angular/compiler';
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import {
   initializeModel,
+  NgDiagramBackgroundComponent,
   NgDiagramComponent,
   provideNgDiagram,
   type NgDiagramConfig,
@@ -11,7 +12,7 @@ import {
 import { CustomNodeComponent } from './node/node.component';
 
 @Component({
-  imports: [NgDiagramComponent],
+  imports: [NgDiagramComponent, NgDiagramBackgroundComponent],
   providers: [provideNgDiagram()],
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
@@ -22,7 +23,9 @@ import { CustomNodeComponent } from './node/node.component';
         [model]="model"
         [config]="config"
         [nodeTemplateMap]="nodeTemplateMap"
-      />
+      >
+        <ng-diagram-background />
+      </ng-diagram>
       <!-- @section-end:usage -->
     </div>
   `,
@@ -38,6 +41,10 @@ export class DiagramComponent {
   config = {
     zoom: {
       max: 3,
+      zoomToFit: {
+        onInit: true,
+        padding: 120,
+      },
     },
   } satisfies NgDiagramConfig;
 
@@ -46,7 +53,7 @@ export class DiagramComponent {
     nodes: [
       {
         id: '1',
-        position: { x: 200, y: 50 },
+        position: { x: 0, y: 0 },
         size: { width: 250, height: 170 },
         autoSize: false,
         // @mark-start